LayoutTests/platform/ios-simulator/editing/shadow/selection-of-shadowroot-expected.txt

CONSOLE MESSAGE: line 3: HierarchyRequestError: DOM Exception 3: A Node was inserted somewhere it doesn't belong.
BEFORE HOST
HOST
AFTER HOST
FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
